+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - cli: Add command `flush` to truncate all short IDs
 - core: Add api `Aim.flush_short_ids`
+- core: Add `DateTimeAnchor::Relative`, remove `DateTimeAnchor::InHours`
 
 ### Changed